move API calls to services, delete utilities
